Use arm64 micromamba for Apple Silicon
Issue:
It is best to use native architecture for the Apple Silicon if avialable.
For soma native binaries, e.g. pytorch, it is critical to use correct architecture, due to the use of architecture specific CPU extensions in the builds

Solution:
Extend download functions for mac to match CPU architecture of VSCode
Update tests to download x64 and arm64
Add downloaded binary architecture check